% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easySenti.R
\name{easySenti}
\alias{easySenti}
\title{easySenti}
\usage{
easySenti(docs, positive, negative, t = 0, sigmoid = FALSE,
  t.s = 0.3, t.s2 = 0.1)
}
\arguments{
\item{positive}{vector. postive words}

\item{negative}{vector. negative words}

\item{t}{numeric. threshold to be positive(and -negative). count of positive - count of negative}

\item{sigmoid}{logic. use or not of sigmoid function}

\item{t.s}{numeric. minimum threshold to be positive. sigmoid of (positive - negative)}

\item{t.s2}{numeric. maximum threshold to be negative. sigmoid of (positive - negative)}

\item{doc}{vector. documents to read}
}
\value{
result vector.
}
\description{
Calculation of sentiment value of documents
}
\examples{
cntSenti(docs = docs, positive=positive, negative=negative, t=0, sigmoid=FALSE, t.s=0.3, t.s2=0.1)
}
